+++ src/Poseidon/CLI/Genoconvert.hs
@@ -23,7 +23,7 @@ import           Pipes                      (MonadIO (liftIO),
 import           Pipes.Safe                 (runSafeT)
 import           SequenceFormats.Eigenstrat (writeEigenstrat)
 import           SequenceFormats.Plink      (writePlink)
-import           System.Directory           (removeFile, doesFileExist)
+import           System.Directory           (removeFile, doesFileExist, createDirectoryIfMissing)
 import           System.FilePath            ((<.>), (</>))
 
 -- | A datatype representing command line options for the validate command
@@ -73,9 +73,13 @@ convertGenoTo outFormat onlyGeno outPath removeOld pac = do
     then logWarning "The genotype data is already in the requested format"
     else do
         -- create new genotype data files
-        let newBaseDir = case outPath of
-                Just x -> x
-                Nothing -> posPacBaseDir pac
+        newBaseDir <- case outPath of
+            Just x -> do
+                -- create new directory
+                logInfo . pack $ "Writing to directory (will be created if missing): " ++ x
+                liftIO $ createDirectoryIfMissing True x
+                return x
+            Nothing -> return $ posPacBaseDir pac
         let [outG, outS, outI] = map (newBaseDir </>) [outGeno, outSnp, outInd]
         anyExists <- or <$> mapM checkFile [outG, outS, outI]
         if anyExists