+++ test/Poseidon/GoldenTestsRunCommands.hs
@@ -4,7 +4,7 @@ module Poseidon.GoldenTestsRunCommands (
     createStaticCheckSumFile, createDynamicCheckSumFile, staticCheckSumFile, dynamicCheckSumFile
     ) where
 
-import           Poseidon.EntitiesList          (readEntitiesFromString, PoseidonEntity(..))
+import           Poseidon.EntitiesList          (readEntitiesFromString, PoseidonEntity(..), EntityInput(..))
 import           Poseidon.CLI.Update            (UpdateOptions (..), runUpdate)
 import           Poseidon.CLI.Genoconvert       (GenoconvertOptions (..), runGenoconvert)
 import           Poseidon.CLI.Init              (InitOptions (..), runInit)
@@ -329,7 +329,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
     -- forge test 1
     let forgeOpts1 = ForgeOptions { 
           _forgeGenoSources  = [PacBaseDir $ testDir </> "Schiffels", PacBaseDir $ testDir </> "Wang"]
-        , _forgeEntitySpec   = Left (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>")
+        , _forgeEntityInput  = [EntitiesDirect (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>")]
         , _forgeSnpFile      = Nothing
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatEigenstrat
@@ -349,7 +349,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
     -- forge test 2
     let forgeOpts2 = ForgeOptions { 
           _forgeGenoSources  = [PacBaseDir $ testDir </> "Schiffels", PacBaseDir $ testDir </> "Wang"]
-        , _forgeEntitySpec   = Left (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>,-<SAMPLE3>")
+        , _forgeEntityInput  = [EntitiesDirect (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>,-<SAMPLE3>")]
         , _forgeSnpFile      = Nothing
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatPlink
@@ -368,7 +368,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
     -- forge test 3
     let forgeOpts3 = ForgeOptions { 
           _forgeGenoSources  = [PacBaseDir $ testDir </> "Schiffels", PacBaseDir $ testDir </> "Wang"]
-        , _forgeEntitySpec   = Right (testEntityFiles </> "goldenTestForgeFile1.txt")
+        , _forgeEntityInput  = [EntitiesFromFile (testEntityFiles </> "goldenTestForgeFile1.txt")]
         , _forgeSnpFile      = Nothing
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatEigenstrat
@@ -390,7 +390,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
     -- forge test 4
     let forgeOpts4 = ForgeOptions { 
           _forgeGenoSources  = [PacBaseDir $ testDir </> "Schiffels", PacBaseDir $ testDir </> "Wang"]
-        , _forgeEntitySpec   = Right (testEntityFiles </> "goldenTestForgeFile2.txt")
+        , _forgeEntityInput  = [EntitiesFromFile (testEntityFiles </> "goldenTestForgeFile2.txt")]
         , _forgeSnpFile      = Nothing
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatPlink
@@ -412,7 +412,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
     -- forge test 5
     let forgeOpts5 = ForgeOptions { 
           _forgeGenoSources  = [PacBaseDir $ testDir </> "Schiffels", PacBaseDir $ testDir </> "Wang"]
-        , _forgeEntitySpec   = Left []
+        , _forgeEntityInput  = []
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatEigenstrat
         , _forgeOutMinimal   = False
@@ -455,7 +455,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
                 , snpSet   = Just SNPSetOther
               }
           ]
-        , _forgeEntitySpec   = Left (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>")
+        , _forgeEntityInput  = [EntitiesDirect (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>")]
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatEigenstrat
         , _forgeOutMinimal   = False
@@ -488,7 +488,7 @@ testPipelineForge testDir checkFilePath testEntityFiles = do
                 , snpSet   = Just SNPSetOther
               }
             ]
-        , _forgeEntitySpec   = Left (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>")
+        , _forgeEntityInput  = [EntitiesDirect (fromRight [] $ readEntitiesFromString "POP2,<SAMPLE2>,<SAMPLE4>")]
         , _forgeIntersect    = False
         , _forgeOutFormat    = GenotypeFormatEigenstrat
         , _forgeOutMinimal   = False
@@ -516,11 +516,9 @@ testPipelineFetch :: FilePath -> FilePath -> IO ()
 testPipelineFetch testDir checkFilePath = do
     let fetchOpts1 = FetchOptions { 
           _jaBaseDirs       = [testDir]
-        , _entityList       = [Pac "2019_Nikitin_LBK"]
-        , _entityFiles      = []
+        , _entityInput      = [EntitiesDirect [Pac "2019_Nikitin_LBK"]]
         , _remoteURL        = "http://c107-224.cloud.gwdg.de:3000"
         , _upgrade          = True
-        , _downloadAllPacs  = False 
         }
     runAndChecksumFiles checkFilePath testDir (usePoseidonLogger NoLog $ runFetch fetchOpts1) "fetch" [
           "2019_Nikitin_LBK" </> "POSEIDON.yml"